chore: Add new packages and update dependencies
This commit updates the `akmods.yml` and `rpm-ostree.yml` files. It adds the `kvmfr` package to the `akmods.yml` file and updates the `rpm-ostree.yml` file to include new packages such as `fish`, `cronie`, `chronie-anacron`, `crontabs`, and `timeshift`. It also removes the `bc` and `borgbackup` packages. These changes aim to improve the functionality and performance of the application.