Remove build-prep.sh script from tacos-os.yml
This commit removes the build-prep.sh script from the tacos-os.yml file. The script is no longer needed as it has been replaced by a more efficient method of preparing the repositories and build environment for the image build process. This change aims to simplify the configuration and improve the overall build process.