chore: Update multimedia and sound-and-video packages
This commit updates the multimedia and sound-and-video packages in the final-steps.sh script. It swaps the ffmpeg-free package with ffmpeg and updates the @multimedia and @sound-and-video package groups. Additionally, it swaps the mesa-va-drivers package with mesa-va-drivers-freeworld and mesa-vdpau-drivers with mesa-vdpau-drivers-freeworld. These changes aim to improve multimedia support and ensure the latest versions of the packages are installed.