chore: Update scripts and enable Howdy
This commit updates the final-steps.sh script to include the installation of Howdy, a facial recognition authentication system. It also adds the necessary steps to enable Howdy in the tacos-os.yml file. By including Howdy, we enhance the security and convenience of the system. Additionally, the remove-fedora-repos.sh script has been added to remove Fedora repositories, improving the overall functionality and maintainability of the codebase.