chore: Update Howdy repository configuration and scripts
This commit updates the enable-howdy-beta-repo.sh script to fix issues with the Howdy repository configuration. It removes unnecessary code related to Google GPG keys and adds the necessary steps to download and install the Google signing keys separately. Additionally, the final-steps.sh script has been modified to use rpm-ostree for Howdy installation. These changes aim to improve the functionality and maintainability of the codebase.